refactor(metrics): group index page into directory
